Platform/ARM/Juno: Remove SSDT UART table
With the introduction of Dynamic AML the DBG2 and SPCR port
generators have been updates to generate the AML code for
describing the Serial ports.

This makes the UART description in the SSDT table redundant.
This patch removes the SsdtUart.asl and makes the corresponding
changes to the configuration manager.

Change-Id: I232108a809b6a69f322acd23f7d2abcd99e3dfc5
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>