Platform/ARM/Morello: Port PCI Express library
Morello platform requires a custom platform-specific PCI Express
library because the native PCI Express Library only allows for a
single ECAM config address to be supplied to it. If there is more
than one PCIe root port, it expects the ECAM regions for all the
root ports to be contiguous. This is not the case with Morello where
the two RPs have their ECAM regions mapped to non-contiguous address
ranges and reside in separate PCIe segments.

This custom plaform-specific PCI Express library, inherited from
MdePkg/BasePciExpressLib, routes the ECAM accesses to the appropriate
ECAM region based on the PCIe segment number in the incoming PCIe
address.

Segment 0 is PCIe root complex and segment 1 is CCIX root complex.

Signed-off-by: Varshit Pandya <varshit.pandya@arm.com>
Change-Id: I624c0aae87efc23d97b269750b0faeaa3b9549ba