AmpereAltraPkg, JadePkg: RTC: Update RTC datetime to SMpro
To get rid of hardware dependency on a chosen RTC chip in the
common SMpro/PMpro firmware, the SMpro will get the RTC datetime from
UEFI through a mailbox interface instead of accessing directly to the
platform RTC chip. This patch is to hook this date configuration to the
runtime RTC library so the SMpro datetime is always updated once having
changes from users both at boot time and runtime.

Note that the CPU performance might be impacted with lack of this
support.

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>