Platform/ARM/Morello: Ensure GOP is enabled at runtime in FVP
Because the FVP do not have any input drivers in UEFI we do not want to
enable the console redirection on it.
However, we want the GOP to be made available to the OS (that may have
drivers).
This is archived by setting a mode if one have not already been set.

Change-Id: Ie36203fedfd531c3acca3a547f085dbf8cdddfc8
Signed-off-by: Varshit Pandya <varshit.pandya@arm.com>