Platform/ARM/Morello: Ensure GOP is enabled at runtime in FVP
Because the FVP do not have any input drivers in UEFI we do not want to
enable the console redirection on it.
However, we want the GOP to be made available to the OS (that may have
drivers).
This is achieved by setting a mode if one has not already been set.

Signed-off-by: Varshit Pandya <varshit.pandya@arm.com>
Change-Id: Ie74240396b14083b82c174cd49c37ebadf6ba941