Platform/ARM/Morello: Fix SMBIOS Type 3 (System Enclosure) length
The SMBIOS header for this table was initialised with the wrong
table's size, which was 3 bytes greater than the correct value.
As a result,  the first three characters from the strings, the
"ARM" in "ARM LTD" for the chassis manufacturer, are erroneously
attributed to the table and so parsers interpret the manufacturer
as " LTD".

Signed-off-by: Jessica Clarke <jessica.clarke@cl.cam.ac.uk>
Signed-off-by: Varshit Pandya <varshit.pandya@arm.com>
Change-Id: I7bd4d4d5b05d6994ae6d937bcf7ca2b14b36a86b