Platform/ARM/Library: Introduce FwsGptSystemFipLib
FwsPlatformLib is platform specific library class used to access the
firmware storage.
In case of Base FVP, this is implemented with FwsGptSystemFipLib which
uses GPT partitions and system fip image.

Following is the PSA-FWU GPT-partition layout for the flash as defined by TF-A:

+----------------------+
|      GPT-HEADER      |
+----------------------+
|    FIP_A (bank0)     |
+----------------------+
|    FIP_B (bank1)     |
+----------------------+
|    FWU-Metadata      |
+----------------------+
|  Bkup-FWU-Metadata   |
+----------------------+

FwsGptSystemFipLib defines API to access the GPT partitions for
updating the images in the banks and to read/update the Metadata.

Signed-off-by: Yeoreum Yun <yeoreum.yun@arm.com>