Platform/Ampere: Switch to generic ResetRuntime DXE driver
The reset runtime DXE driver is deprecated and will be removed soon. It
is superseded by a generic implementation in MdeModulePkg, which is
shared between all architectures, and implements the notification
protocols that the EFI spec describes. So move the Ampere Jade platform
to this implementation the of ResetSystem EFI runtime service.

Reviewed-by: Nhi Pham <nhi@os.amperecomputing.com>
Tested-by: Nhi Pham <nhi@os.amperecomputing.com>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>