Platform/SbsaQemu: Switch to generic ArmPsciResetSystemLib
The PSCI reset libraries in ArmPkg are being consolidated into a single
one to reduce the maintenance burden of having multiple libraries doing
the same thing in a slightly different way.

Move this platform to the generic ArmPsciResetSystemLib, and drop the
dependency on ArmSmcPsciResetSystemLib, which is deprecated and will be
removed.

Cc: Marcin Juszkiewicz <marcin.juszkiewicz@linaro.org>
Cc: Graeme Gregory <graeme@xora.org.uk>
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>