Silicon/Ampere: Switch to generic ArmPsciResetSystemLib
The PSCI reset libraries in ArmPkg are being consolidated into a single
one to reduce the maintenance burden of having multiple libraries doing
the same thing in a slightly different way.

Move this platform to the generic ArmPsciResetSystemLib, and drop the
dependency on ArmSmcPsciResetSystemLib, which is deprecated and will be
removed.

Cc: Nhi Pham <nhi@os.amperecomputing.com>
Cc: Chuong Tran <chuong@os.amperecomputing.com>
Cc: Rebecca Cran <rebecca@os.amperecomputing.com>
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>
Reviewed-by: Nhi Pham <nhi@os.amperecomputing.com>