Marvell/Odyssey: Switch to generic ResetRuntime DXE driver
The reset runtime DXE driver is deprecated and was recently removed. It
is superseded by a generic implementation in MdeModulePkg, which is
shared between all architectures, and implements the notification
protocols that the EFI spec describes. So move the Marvell Odyssey platform
to this implementation the of ResetSystem EFI runtime service.

Signed-off-by: Marcin Wojtas <marcin.s.wojtas@gmail.com>