Platform/Intel: Override BaseMemoryLib if SMM profile enable
The instructions in BaseMemoryLibRepStr could potentially
trigger multiple #PF (Page Faults) before the Debug Trap
handler is invoked if the size of the memory being copied
exceeds 8KB. This would result in the number of recorded
old entries bigger than the MAX_PF_ENTRY_COUNT limit of
10. The relevant instructions are:
    mov     rcx, r8
    and     r8, 7
    shr     rcx, 3
    rep     movsq

To prevent this issue, override the BaseMemoryLib with
BaseMemoryLibSse2 when the SMM profile is enabled. This
change avoids the use of the "rep movsq" instruction, which
is the source of the problem.

Signed-off-by: Jiaxin Wu <jiaxin.wu@intel.com>