ManageabilityPkg: fix resource cleanup
This patch corrects the order of operations when freeing resources.

Also this patch fixes possible access to NULL pointer in
ReleaseTransportSession(). If NULL pointer is passed to this function,
the container_of(NULL) will return some invalid pointer.

The behavior of this function remains the same. If some valid pointer
is passed it will be dereferenced and freed. Function returns success
only if pointer to the passed token is equal to the saved
mSingleSessionToken value.

Cc: Abner Chang <abner.chang@amd.com>
Cc: Abdul Lateef Attar <AbdulLateef.Attar@amd.com>
Cc: Nickle Wang <nicklew@nvidia.com>
Cc: Nhi Pham <nhi@os.amperecomputing.com>
Signed-off-by: Mike Maslenkin <mike.maslenkin@gmail.com>