AmpereAltraPkg: Add PlatformInitDxe module
In the FailSafe context, there is a field that indicates which setting
is being used to boot (BOOT_LAST_KNOWN_SETTINGS, BOOT_DEFAULT_SETTINGS,
BOOT_NORMAL). On the SCP and TF-A side, they will check their NVPARAM
for FailSafe (NV_SI_PMPRO_FAILURE_FAILSAFE - NV_SI_ATF_FAILURE_FAILSAFE)
to determine which setting to use. If a FailSafe event occurs at SCP or
TF-A, it is the responsibility of UEFI during the DXE phase to clear the
FailSafe context to ensure normal behavior.

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>