AmpereAltraPkg: Add TPM2 support
In Ampere Altra, TPM is a secured device and can only be accessed
in S-EL0 secure partition. Upper layers like UEFI and Operating Systems
need the intermediate Secured Partition Manager (SPM) in Trusted
Firmware BL31 to establish a communication with the TPM device which is
physically connected to the Socket 0 SPI bus 1.

The communication interface, designed as TPM CRB buffer, is inline with
the TPM2.0 Mobile Command Response Buffer Interface specification.

This patch introduces TPM2 support for Ampere Altra Mt. Jade platform
based on edk2 implementation. Due to differences in platform design, the
TPM2 code has been forked from the edk2/SecurityPkg and adapted to meet
our specific requirements. The implementation focuses on meeting the
mandatory TPM2 functionality and does not include all TPM2 or PPI
features, limiting the scope to what is essential for Ampere platforms.
We may expand the implementation later if more features are needed.

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>