IntelSiliconPkg/SmmAccess: Fix potential integer overflow in loop comparison
SmmAccessDxe:

`mSmmAccess.NumberRegions` is a `UINTN` while the loop index variable
compared against it is a `UINT8`. This can lead to an overflow of the
loop index for `mSmmAccess.NumberRegions` for values larger than
`UINT8_MAX`. This change makes `Index` a `UINTN` to match in width.

PeiSmmAccessLib:

Fixes a similar `UINT8` loop index issue.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>