Ampere: Add full set of I2C buses to Ac01.h
Boards based on the Altra SoC from vendors such as ADLINK are
configured differently and require access to I2C buses not
listed in Ac01.h. Fix this by expanding AC01_I2C_BASE_ADDRESS_LIST to
contain the full set of buses available.

Since this changes the indexing of the buses, update the Mt Jade files
to match. Add a new PCD to contain the bus number of the RTC since this
is different between Mt Jade and ADLINK's COM-HPC-ALT platform.

Signed-off-by: Rebecca Cran <rebecca@os.amperecomputing.com>