JadePkg: Add PlatformBootManagerDxe to customize boot options
This introduces a PlatformBootManagerDxe driver, which implements the
EDKII_PLATFORM_BOOT_MANAGER_PROTOCOL to enhance the presentation of boot
options for the platform.

The primary movitation behind this change is to provide a user-friendly
boot experience by customizing the list of boot options generated by
EDK2 boot manager. In addition, the driver filters out non-OS entries.
This reduces unnecessary boot options, ensuring the boot menu is cleaner
and less confusing.

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>