Pkg-Module: Features/VirtualKeyboardDxe PART 2
The Keyboard.c driver did not include ScanCodes like F1,F2,F12,etc.
and UnicodeChars like Backspace, Tab in VK_NOTIFY NotifyList
which is why these keys were not getting registered and
not getting pushed in the KeyBuffer stack which resulted
in undefined behavior of continuous backspace and not recognizing the
relevant key push after the buffer overflow condition.

Signed-off-by: Darbin Reyes <darbin.reyes@intel.com>