Platform/Sgi: Add ACPI tables for RD-V3-Cfg1 platform
Arm's RD-V3-Cfg1 platform is a variant of the RD-V3 platform. Compared
with RD-V3 platform it has a reduced core count of 8 Neoverse-V3 CPUs
with an interconnect of size 3x3. There is one CPU per cluster in the
system, so it has a total of 8 clusters. To prepare for supporting this
platform, add the initial set of ACPI tables and reuse existing ACPI
tables where applicable to boot an operating system on this platform.

Signed-off-by: Dhinesh Balu <dhinesh.balu@arm.com>