.github: Add EDK II Code First form
The EDK II Code First process is being revised to align with GitHub.

The actual process changes will be made to the process page on the
TianoCore wiki:

https://github.com/tianocore/tianocore.github.io/wiki/EDK-II-Code-First-Process

These changes:

- Remove the "code first" checkbox from existing forms.
- Add a "code first" issue form that is used to submit issues
tracking code first changes.

The primary template filled in for code first changes is included
with no change from the prior template.

Note: This change is made in edk2-platforms so its issue submission
process is consistent with edk2 but most code first items are
expected to use an edk2 repo issue as their primary tracking issue.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>