VExpressPkg: Install protocol after initializing repository in CfgMgr
The configuration manager repository should be initialized first
before installing protocols. The current implementation does the
opposite. If the repo initialization fails after the protocol
is already installed and the protocol is accessed without a
backing repo, this would result in crashes.

This fix makes sure the correct order is used to avoid such
issues

Signed-off-by: Mohamed Gamal Morsy <mohamed.morsy@arm.com>