AmpereAltraPkg/MemInfoDxe: Fix ECC Mode in Memory Formset
The ECC Mode in Memory Configuration Formset is not working properly
because "Auto" mode is absent. It causes an critical issue that DDR
always runs with Non-ECC mode a.k.a "Disabled" mode. This patch fixes
that issue by adding "Auto" Option to "ECC Mode" configuration that
aligns with the original design.

This patch also updates Platform UUID to clean up UEFI NVRAM region
and restore the settings to default value.

Signed-off-by: Chuong Tran <chuong@os.amperecomputing.com>