Platform/SiFive/U5SeriesPkg: Add modified RamDisk driver
Since there is no storage driver in EDK2 for U540, add a modified
RamDisk driver where qemu can load the linux+rootfs at a known address
and a RamDisk is automatically created at this address. EDK2 can load
linux from this RamDisk and mount the initramfs as rootfs.

This is only for qemu sifive_u machine. It is not required if someone
adds a proper storage driver.

Signed-off-by: Tuan Phan <tphan@ventanamicro.com>
Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>