Silicon/Ampere: Improve SLC cache and ANC strings
Improve some HII strings related to ANC and SLC cache.

Clarify in the help string for ANC mode that 'ANC' means
Ampere NUMA Control. And simplify the rest of the string
to avoid repetition.

Replace '$' with the word 'Cache'. While '$' might be a common
shorthand in the industry, it confuses other people.

In the platform board information page, the help string was wrong:
it doesn't indicate SLC as L3 Cache, but simply the amount of
System Level Cache.

Signed-off-by: Rebecca Cran <rebecca@bsdio.com>