Silicon/Qualcomm: Add support for UFS driver
UFS controller on Qcom SoCs is attached to a non-discoverable bus, so it
is registered as a non-discoverable device with DMA being non-cache
coherent.

This UFS driver assumes that the prior stage XBL on Qcom SoCs has fully
initialized the UFS controller where the primary boot medium is UFS flash
on Qcom SoCs. Given that the initialization is skipped here and only the
UFS controller MMIO address space is registered for UFS I/O to take place.

Signed-off-by: Sumit Garg <sumit.garg@oss.qualcomm.com>