Drivers/OptionRomPkg: Remove gEfiScsiPassThruProtocolGuid
The legacy ScsiPassThruProtocol has been removed as per
the UEFI 2.10A specification.
This commit, removes references to ScsiPassThruProtocol and
transitions the implementation to only use ExtScsiPassThruProtocol,
which provides enhanced support for device enumeration,
target/LUN discovery, and broader SCSI compatibility.
Removed PcdSupportExtScsiPassThru and PcdSupportScsiPassThru
as well, because the only available interface is ExtScsiPassThru.

Cc: Sachin Ganesh <sachinganesh@ami.com>
Signed-off-by: Gowtham M <gowthamm@ami.com>